Extract GitHub DTO types
